[ARM] distribute shifts as muls
