+++ src/CMakeLists.txt
@@ -57,6 +57,7 @@ set(HEADER_FILES
     DeviceArgument.h
     DeviceInterface.h
     Dimension.h
+    DistributeShifts.h
     EarlyFree.h
     Elf.h
     EliminateBoolVectors.h
@@ -227,6 +228,7 @@ set(SOURCE_FILES
     DeviceArgument.cpp
     DeviceInterface.cpp
     Dimension.cpp
+    DistributeShifts.cpp
     EarlyFree.cpp
     Elf.cpp
     EliminateBoolVectors.cpp
